requirements.txt
langchain>=0.1.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langgraph>=0.1.0,<0.3.0
streamlit>=1.28.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
typing-extensions>=4.0.0,<5.0.0
pydantic>=2.0.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
plotly>=5.15.0,<6.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0
numpy>=1.24.0,<2.0.0
